requirements.txt
ipython==8.27.0
openai==1.52.0
together==1.3.3
python-dotenv==1.0.1
requests==2.32.3
matplotlib==3.9.2
numpy==1.26.2
pytest==8.3.3
pytest-cov==5.0.0
flake8==7.1.1
pylint==3.3.1
codespell==2.3.0
autoflake==2.3.1
anthropic==0.37.1
boto3==1.35.49
moto==5.0.18
groq==0.11.0